unify api_utils and lambda_util to api_utils
